% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTMOST_fa.R
\name{cv_glasso}
\alias{cv_glasso}
\title{cv_glasso}
\usage{
cv_glasso(
  X_all,
  Y_all,
  X_test,
  Y_test,
  foldid,
  Xnorm,
  bannot,
  nlambda = 10,
  nalpha = 4,
  ...
)
}
\arguments{
\item{X_all}{C-list of N_c * M matrix of genotypes for training.}

\item{Y_all}{C-list of N_c vector of molecular phenotypes for training.}

\item{X_test}{C-list of N_c * M matrix of genotypes for testing.}

\item{Y_test}{C-list of N_c vector of molecular phenotypes for testing.}

\item{foldid}{N vector of fold IDs for k-fold cross-validation. The names of the vector are the individual IDs.}

\item{Xnorm}{M vector of the genotypic variances.}

\item{bannot}{M vector of the genotypic annotations. The values should be within: 1: Active; 0: Nuetral; -1: Inactive.}

\item{nlambda}{the number of tuning parameters lambda.}

\item{nalpha}{the number of tuning parameters alpha.}

\item{...}{additional arguments.}
}
\value{
A \code{"UTMOST_fa"} object with elements for the training model and testing results:
}
\description{
Perform k-fold cross-validation to train a
}
\details{
TBD
}
\references{
TBD
}
\author{
Chen Lin, Hongyu Zhao
}
